Added the start of a package system.
As you can probably tell, this completely broke macros as well as some
other code. I put a bunch of band-aids on, such as using cl-intern in
a bunch of places that it definitely should not be. I'm going to start
working on a way to fix the problem. One solution that may work is
creating a new versions of quote and quasiquote that will use
cl-intern.