Implemented packages as actual objects and moved around code.
This allows for a couple small changes. Since a name can now be
associated with a package, symbols can now be prefixed by the
package name when they are printed.

I moved around the class definitions to the top of the file to avoid
any compiler warnings. I plan on eventually moving them to a seperate
file altogether once the interpreter gets large enough.