Made some adjustments to the package system.
Now it is possible to create multiple packages. The ^ reader will now
always intern in the the interpreter's CL pakcage. The are some
problems with typing in a symbol from another package, as it is caught
by the implementation CL. I plan on getting around this problem by
starting the reader for Clint.

Next I am going to reorganize the code and add tests. Then I am going
to finish off the package system and start working on the reader.