Added an implementation for cons pairs.
It is unclear how this is going to work in with the rest of Clint, so
I am leaving it unattached right now. I might never actually use them
and just use the ICL lists instead. If I decide to use them, I'm going
to have to rewrite all of the code so far to use cl-cons pairs instead
of the ICL cons pairs.